+++ src/main/java/Duke/Storage.java
@@ -1,3 +1,6 @@
+package Duke;
+
+import Task.Task;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.FileWriter;
