+++ src/main/java/Duke/TaskManager.java
@@ -1,3 +1,10 @@
+package Duke;
+
+import Task.Deadline;
+import Task.Event;
+import Task.Task;
+import Task.Todo;
+
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.ArrayList;
